full typescript development
