--- .vscode/extensions.json
@@ -1,13 +1,0 @@
-{
-  "recommendations": [
-    "aaron-bond.better-comments",
-    "dbaeumer.vscode-eslint",
-    "eamodio.gitlens",
-    "editorconfig.editorconfig",
-    "visualstudioexptteam.vscodeintellicode",
-    "fabiospampinato.vscode-todo-plus",
-    "christian-kohler.path-intellisense",
-    "naumovs.color-highlight",
-    "coenraads.bracket-pair-colorizer-2"
-  ]
-}
