--- .vscode/settings.json
@@ -1,15 +1,0 @@
-{
-  "editor.formatOnSave": true,
-  "eslint.format.enable": true,
-  "eslint.codeActionsOnSave.mode": "all",
-  "eslint.enable": true,
-  "[javascript]": {
-    "editor.defaultFormatter": "dbaeumer.vscode-eslint"
-  },
-  "[javascriptreact]": {
-    "editor.defaultFormatter": "dbaeumer.vscode-eslint"
-  },
-  "emmet.includeLanguages": {
-    "javascript": "javascriptreact"
-  }
-}
