+++ package.json
@@ -1,69 +1,64 @@
 {
   "name": "2do",
   "version": "0.1.0",
-  "main": "public/electron.js",
-  "description": "A complete todo app for you or your company",
   "author": "2DON (https://github.com/2DON)",
+  "description": "project management made easy",
+  "homepage": ".",
+  "main": "build/electron/main.js",
   "private": true,
   "scripts": {
-    "dev": "concurrently \"yarn react:dev\" \"wait-on http://localhost:3000 && yarn electron:dev\"",
-    "react:dev": "cross-env BROWSER=none react-scripts start",
-    "prereact:build": "yarn run lint:fix",
-    "react:build": "cross-env GENERATE_SOURCEMAP=false react-scripts build",
-    "electron:dev": "electron .",
-    "electron:build": "electron-builder",
-    "prebuild": "yarn run react:build",
+    "postinstall": "electron-builder install-app-deps",
+    "dev": "concurrently \"cross-env BROWSER=none yarn react:dev\" \"wait-on http://localhost:3000 && yarn run electron:dev\"",
+    "react:dev": "react-scripts start",
+    "electron:dev": "yarn run electron:build && electron .",
     "build": "yarn run electron:build",
-    "lint": "eslint . --ignore-path .gitignore --ext .js,.jsx,ts,tsx --cache",
-    "lint:fix": "yarn run lint --fix",
-    "postlint:fix": "prettier --ignore-path .gitignore --single-quote --write ./**/*.{md,mdx,js,jsx,ts,tsx,json,html,css,less,scss,yaml,yml}"
+    "prebuild": "yarn run react:build",
+    "react:build": "cross-env GENERATE_SOURCEMAP=false react-scripts build",
+    "electron:build": "tsc -p electron",
+    "package": "electron-builder",
+    "prepackage": "yarn run build",
+    "test": "react-scripts test"
   },
   "dependencies": {
+    "@types/electron-devtools-installer": "^2.2.0",
+    "@types/jest": "26.0.14",
+    "@types/node": "14.11.10",
+    "@types/react": "16.9.53",
+    "@types/react-dom": "16.9.8",
+    "electron-devtools-installer": "^3.1.1",
     "electron-is-dev": "^1.2.0",
     "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
     "react-router-dom": "^5.2.0",
+    "react-scripts": "3.4.3",
+    "typescript": "4.0.3",
     "web-vitals": "^0.2.4"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
     "@types/prop-types": "^15.7.3",
-    "@types/react": "^16.9.53",
-    "@types/react-dom": "^16.9.8",
     "@types/react-router-dom": "^5.1.6",
     "concurrently": "^5.3.0",
     "cross-env": "^7.0.2",
-    "electron": "^10.1.5",
+    "electron": "^10.1.3",
     "electron-builder": "^22.9.1",
     "husky": "^4.3.0",
     "lint-staged": "^10.5.0",
     "node-sass": "^4.14.1",
-    "react-scripts": "4.0.0",
-    "typescript": "^4.0.3",
     "wait-on": "^5.2.0"
   },
   "browserslist": "last 1 electron version",
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "prettier": "@imaginary-cloud/prettier-config",
-  "homepage": "./",
   "build": {
     "appId": "io.github.2don.2do-electron",
+    "extends": null,
     "files": [
-      "build/**/*",
-      "node_modules/**/*"
+      "build/**/*"
     ],
     "directories": {
       "buildResources": "assets"
     }
+  },
+  "eslintConfig": {
+    "extends": "react-app"
   }
 }
